lab2/requirements.txt
environs
semantic_version
pytest
starlette_exporter
fastapi
uvicorn
pydantic
requests